requirements.txt
streamlit
torch
torchvision
Pillow
matplotlib
numpy
seaborn
scikit-learn
opencv-python-headless